import feature 0.9

+++ include/str_op.h
@@ -84,8 +84,8 @@ int check_array_input(char *value);
 int count_delim(char *delim, char *str);
 int find_double_delim(char *delim, char *str, int *pos);
 int find_delim_in_fields(char *delim, char *str, int *pos);
-void pack(uint32_t n, uint16_t *digits_indexes);
-long long unpack(int *digits_index);
-void print_pack_str(int *digits_index);
+void pack(uint32_t n, uint8_t *digits_indexes);
+long long unpack(uint8_t *digits_index);
+void print_pack_str(uint8_t *digits_index);
 
 #endif /* STR_OP_H */
